implement a bunch of equations incorrectly - dydw needs to take into account all the weights and inputs when determining if it passes the relu threshold, not just the ones affected by the partial derivative

+++ hybrid_output.py
@@ -1,7 +1,8 @@
 import os
 line = "========================================"
 block = '<>'
-os.remove("out.txt")
+if os.path.isfile("out.txt"):
+    os.remove("out.txt")
 def open_log():
     return open("out.txt", 'a')
